requirements.txt
asgiref==3.8.1
blinker==1.9.0
chardet==5.2.0
click==8.1.8
Django==5.1.5
Flask==3.1.0
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
reportlab==4.2.5
sqlparse==0.5.3
Werkzeug==3.1.3
